new grid feature
An event listener was added to the NEW GRID button, and when it is clicked the user is supposed to enter a new value for the dimensions of their grid, and that new grid is supposed to be created with the for loop and displayed on the screen. However, I am running into some issues. The area of the boxes are stil the same and the number of total boxes is different than expeceted.