Updated NEW GRID feature
I altered the selectors of the html page by adjusting the div element in the html page to just be a container for another container holding the original and new grids. So when the user clicks in the NEW GRID button, a function will be executed which removes the original child selector acting as a container for the grid, and add a new child element which is a container for the new grid.